requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# AI/ML dependencies
openai>=1.3.0
transformers>=4.35.0
torch>=2.1.0
langchain>=0.0.340
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Speech processing
speechrecognition>=3.10.0
pydub>=0.25.1
whisper>=1.1.10

# Translation
googletrans==4.0.0rc1

# Database
sqlite3
pymongo>=4.6.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.1.0

# Additional dependencies
httpcore==0.18.0
httpx==0.25.0
